build(secrets): centralize validation and modernize CI bypass
Centralize the secrets.properties enforcement logic into a single
authoritative check in the root build.gradle.kts file, removing
heavy duplication across six subprojects.

Replace brittle runtime environment checks with safe, build-time
BuildConfig.IS_CI constants for automated testing pipelines.

- Create root build.gradle.kts for workspace-wide validation
- Remove duplicated secrets-checking blocks from subproject scripts
- Inject BuildConfig.IS_CI and apply manifest placeholders in CI
- Update Application classes to respect BuildConfig.IS_CI early returns
- Untrack LOG.md and add to local exclude file
- Fill out missing PLACES_API_KEY in defaults properties
