--- LOG.md
@@ -1,8 +1,0 @@
-
-### Task: Implement Monster Tour and Long Press popup (Project: Maps3D Samples)
-* **Start:** [Unknown] (Branch: main)
-* **End:** $(date +"%I:%M:%S %p %Z")
-* **Time Spent:** [Unknown]
-* **Purpose:** The user wanted to automate the traversal of the various monster markers in the map, and add a quick-select popup menu to the random button, configuring properties like altitude mode via JSON.
-* **Summary of Work:** Shifted marker definitions to a central `monsters.json` data file. Implemented an asynchronous touring sequence in both Kotlin (Coroutines) and Java (Handlers) to fly to, orbit, and display info popovers for each character. Wired up a standard `android.widget.PopupMenu` for the random button long-press. Adjusted altitude modes in JSON.
-* **Status:** **[IN PROGRESS (Unmerged)]**
