+++ Maps3DSamples/ApiDemos/kotlin-app/src/main/java/com/example/maps3dkotlin/Maps3DKotlinApplication.kt
@@ -49,6 +49,9 @@ class Maps3DKotlinApplication : Application() {
      * incorrectly configured, and a RuntimeException is thrown.
      */
     private fun checkApiKey() {
+        if (BuildConfig.IS_CI) {
+            return
+        }
         try {
             val appInfo =
                 packageManager.getApplicationInfo(packageName, PackageManager.GET_META_DATA)
