+++ Maps3DSamples/advanced/app/build.gradle.kts
@@ -14,63 +14,7 @@
  * limitations under the License.
  */
 
-import java.util.Properties
-import org.gradle.api.GradleException
-
-// Check for secrets.properties file and valid API key before proceeding with build tasks.
-val secretsFile = rootProject.file("secrets.properties")
-val isCI = System.getenv("CI")?.toBoolean() ?: false
-
-if (!isCI) {
-    val requestedTasks = gradle.startParameter.taskNames
-    if (requestedTasks.isEmpty() && !secretsFile.exists()) {
-        // It's likely an IDE sync if no tasks are specified, so just issue a warning.
-        println("Warning: secrets.properties not found. Gradle sync may succeed, but building/running the app will fail.")
-    } else if (requestedTasks.isNotEmpty()) {
-        val buildTaskKeywords = listOf("build", "install", "assemble")
-        val isBuildTask = requestedTasks.any { task ->
-            buildTaskKeywords.any { keyword ->
-                task.contains(keyword, ignoreCase = true)
-            }
-        }
-
-        val testTaskKeywords = listOf("test", "report", "lint")
-        val isTestTask = requestedTasks.any { task ->
-            testTaskKeywords.any { keyword ->
-                task.contains(keyword, ignoreCase = true)
-            }
-        }
-
-        val isDebugTask = requestedTasks.any { task ->
-            task.contains("Debug", ignoreCase = true) || task.contains("installAndLaunch", ignoreCase = true)
-        }
-
-        if (isBuildTask && !isTestTask && isDebugTask) {
-            val defaultsFile = rootProject.file("local.defaults.properties")
-            val requiredKeysMessage = if (defaultsFile.exists()) {
-                defaultsFile.readText()
-            } else {
-                "MAPS3D_API_KEY=<YOUR_API_KEY>"
-            }
-
-            if (!secretsFile.exists()) {
-                throw GradleException("secrets.properties file not found. Please create a 'secrets.properties' file in the root project directory with the following content:\n\n$requiredKeysMessage")
-            }
-
-            val secrets = Properties()
-            secretsFile.inputStream().use { secrets.load(it) }
-            val apiKey = secrets.getProperty("MAPS3D_API_KEY")
-
-            if (apiKey.isNullOrBlank() || !apiKey.matches(Regex("^AIza[a-zA-Z0-9_-]{35}$"))) {
-                throw GradleException("Invalid or missing MAPS3D_API_KEY in secrets.properties. Please provide a valid Google Maps API key (starts with 'AIza').")
-            }
-
-            if (secrets.getProperty("MAPS_API_KEY") != null) {
-                println("Warning: Found MAPS_API_KEY in secrets.properties. This project relies exclusively on MAPS3D_API_KEY.")
-            }
-        }
-    }
-}
+val isCI = rootProject.extra["isCI"] as? Boolean ?: false
 
 plugins {
     alias(libs.plugins.android.application)
@@ -98,6 +42,11 @@ android {
         versionName = "1.7.0"
 
         testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
+
+        if (isCI) {
+            manifestPlaceholders["MAPS3D_API_KEY"] = "DEFAULT_API_KEY"
+            manifestPlaceholders["PLACES_API_KEY"] = "DEFAULT_API_KEY"
+        }
     }
 
     buildTypes {
