+++ local.defaults.properties
@@ -1,1 +1,2 @@
 MAPS3D_API_KEY=DEFAULT_API_KEY
+PLACES_API_KEY=DEFAULT_API_KEY
